Pdf is read-only after first save_as call

+++ lib/pdf.rb
@@ -8,6 +8,7 @@ require 'stringio'
 # PDF Document with a form that can be read, filled, and saved.
 class Pdf
   class NonexistentFieldError < Exception; end
+  class ReadOnlyError < Exception; end
   include_package "com.itextpdf.text.pdf"
   include_package "java.io"
 
@@ -16,15 +17,21 @@ class Pdf
     @output_buffer = StringIO.new
     reader = PdfReader.new(@data.to_java_bytes)
     @stamper = PdfStamper.new(reader, @output_buffer.to_outputstream)
+    @saved = false
   end
 
   # Writes PDF to +path+. If +flatten+ is true, also flattens the form
   # so that the form is printed on the PDF document but the form is no
   # longer editable.
   #
+  # Once a this has been called the PDF becomes read-only and any
+  # subsequent calls to +save_as+, +set_field+, or +set_fields+
+  # will raise an ReadOnlyError.
+  #
   # save_as returns *UNTESTED* if the PDF form is not valid
-  # save_as returns *UNTESTED* if the file cannot be written
   def save_as(path, flatten=false)
+    raise ReadOnlyError.new("Cannot save a previously saved pdf") if @saved
+    @saved = true
     @stamper.setFormFlattening(flatten)
     @stamper.close
     File.open(path, 'wb') do |file|
@@ -33,7 +40,7 @@ class Pdf
   end
 
   # Returns fields defined in this PDF form and their values, if any.
-  # fields returns *UNTESTED* if PDF document does not contain a form
+  # fields returns an empty hash if PDF document does not contain a form
   def fields
     form = @stamper.getAcroFields
     h = {}
@@ -47,7 +54,6 @@ class Pdf
   #
   # Raises Pdf::NonexistentFieldError if field does not exist.
   # +name+:: Symbol name of field to retrieve
-  # get_field returns *UNTESTED* if PDF document does not contain a form
   def get_field(name)
     raise NonexistentFieldError.new("'#{name}' field does not exist in form") unless has_field?(name)
     @stamper.getAcroFields.getField(name.to_s)
@@ -58,6 +64,7 @@ class Pdf
   # Raises Pdf::NonexistentFieldError if field does not exist.
   # +name+: Symbol naming the field to write
   def set_field(name, value)
+    raise ReadOnlyError.new('Previously saved pdfs are read-only') if @saved
     name = name.to_sym
     raise NonexistentFieldError.new("'#{name}' field does not exist in form") unless has_field?(name)
     @stamper.getAcroFields.setField(name.to_s, value)
@@ -70,7 +77,6 @@ class Pdf
   # Raises Pdf::NonexistentFieldError if any field does not exist.
   # +fields+:: A hash of :key => "value" pairs.
   def set_fields(fields)
-    form = @stamper.getAcroFields
     fields.each_pair do |name, value|
       set_field(name, value)
     end
@@ -84,13 +90,8 @@ class Pdf
     fields.key?(name.to_sym)
   end
 
+  # true if the receiving Pdf instance has a form
   def has_form?
     @stamper.getAcroFields.getFields.size > 0
   end
-
-  # TODO: def has_form?, returns true if PDF has an AcroFields object
-
-  # BUG: set_field does not validate field on PDF form. The fix is to
-  # keep the PDF in an in-memory buffer and use PdfStamper to write
-  # immediately to it rather than deferring changes.
 end
